TextTree: Switch to normal Int64s instead of Numerics

+++ Com.Halfdecent.TextTree/DataToken.cs
@@ -1,5 +1,5 @@
 // -----------------------------------------------------------------------------
-// Copyright (c) 2010, 2011, 2012
+// Copyright (c) 2010, 2011, 2012, 2013
 // Ron MacNeil <macro187 AT users DOT sourceforge DOT net>
 //
 // Permission to use, copy, modify, and distribute this software for any
@@ -44,7 +44,7 @@ DataToken
 public
 DataToken(
     string data,
-    int    lineNumber
+    long   lineNumber
 )
     : base( lineNumber )
 {
